requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
geoalchemy2>=0.12.5
celery>=5.0.0
redis>=4.0.0
python-multipart>=0.0.5
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
pydantic[email]>=2.0.0
geojson-pydantic>=1.0.0
geojson>=3.0.0
python-jose[cryptography]>=3.3.0
pillow>=11.1.0
exifread>=2.3.2
alembic>=1.7.0
minio==7.2.0
shapely>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
geopandas>=0.13.0
rasterio>=1.3.0
pyproj>=3.5.0
matplotlib>=3.0.0
fastapi-mail>=1.4.1
torch>=2.0.0
torchvision>=0.15.0
